Tying up loose ends: styling, bug fixes, edge cases...also removing old box menu
