+++ assets/styles.css
@@ -248,8 +248,7 @@ select {
   display: flex;
   flex-direction: column;
 
-  border-bottom: solid black 2px;
-  padding-bottom: 20px;
+  padding-bottom: 5px;
 }
 
 #enterDialogue, #enterName {
@@ -258,10 +257,13 @@ select {
   align-items: flex-start;
 }
 
-#dialogueHeader {
-  margin-top: 25px;
+#dialogueHeader, #nameHeader {
   font-size: 12pt;
   text-align: left;
+}
+
+#nameHeader {
+  margin-top: 25px;
 }
 
 #textField, #nameField {
@@ -277,23 +279,32 @@ select {
   box-shadow: 4px 4px black;
 }
 
+#downloadAndUpload {
+  display: flex;
+  justify-content: space-between;
+  border-bottom: 2px solid black ;
+  padding-bottom: 20px;
+}
+
 #download {
   padding-top: 5px;
   padding-bottom: 5px;
-  width: 99.5%;
+  width: 50%;
+  margin-right: 30px;
 
   display: flex;
   align-items: center;
+  justify-content: center;
 
   border: 2px solid black;
   box-shadow: 3px 3px black;
   background-color: white;
-  font-size: 18pt;
+  font-size: 14pt;
   color: black;
   text-align: center;
 }
 
-#download:hover {
+#download:hover, #upload:hover {
   border: 2px solid white;
   box-shadow: 3px 3px white;
   background-color: black;
@@ -301,8 +312,30 @@ select {
   fill: white;
 }
 
-svg {
-  padding-left: 115px;
+#upload {
+  width: 50%;
+  padding-top: 5px;
+  padding-bottom: 5px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  border: 2px solid black;
+  box-shadow: 3px 3px black;
+  background-color: white;
+  font-size: 14pt;
+  color: black;
+  text-align: center;
+}
+
+#uploadSizeMessage {
+  font-size: 8pt;
+}
+
+#hiddenUpload {
+  display: none;
 }
 
 #genLinks {
@@ -402,7 +435,7 @@ svg {
 #vanillaFont {
   margin-right: 0.5em;
   padding: 0.5em;
-  padding-top: 1.1em;
+  padding-top: 0.6em;
   font-size: 10pt;
 }
 
@@ -410,7 +443,7 @@ svg {
   margin-right: 0.5em;
   margin-left: 0.5em;
   padding: 0.5em;
-  padding-top: 1.1em;
+  padding-top: 0.6em;
   font-family: 'Optima nova LT';
   font-size: 10pt;
 }
@@ -418,6 +451,7 @@ svg {
 #japaneseFont {
   margin-left: 0.5em;
   padding: 0.5em;
+  padding-top: 0.6em;
   font-family: 'SlumpDB';
   font-size: 10pt;
 }
@@ -425,6 +459,7 @@ svg {
 #koreanFont {
   margin-left: 1em;
   padding: 0.5em;
+  padding-top: 0.6em;
   font-family: 'aCinema';
   font-size: 10pt;
 }
