+++ components/Menus.tsx
@@ -2,12 +2,9 @@ import { useState, useEffect } from 'react';
 import CharacterMenu from './CharacterMenu';
 import EmotionMenu from './EmotionMenu';
 import CostumeMenu from './CostumeMenu';
-import BoxMenu from './BoxMenu';
-
-import { boxDirectory } from '../utils/boxFinder'
 
 const Menus = ({ char, setChar, emote, setEmote, costume, 
-  setCostume, setPortrait, font, box, boxFont, setBox, selection, setSelection }) => {
+  setCostume, setPortrait, font, box, name, setName, setBox }) => {
   const [emoteMenus, setEmoteMenus] = useState<string[]>([]);
   const [costumeMenus, setCostumeMenus] = useState<string[]>([]);
 
@@ -22,22 +19,17 @@ const Menus = ({ char, setChar, emote, setEmote, costume,
     setEmoteMenus,
     costumeMenus,
     setCostumeMenus,
+    name,
+    setName,
     font,
     box,
     setBox,
-    selection,
-    setSelection,
   };
 
   useEffect(() => {
-    // setBox(`../images/boxes/db-${selection}-${boxFont}.png`);
     setPortrait(`../images/portraits/${char}/${emote}/${char}-${emote}-${costume}.png`);
     return;
-  }, [char, emote, costume, selection]);
-
-  useEffect(() => {
-    setSelection(char);
-  }, [char]);
+  }, [char, emote, costume]);
 
   return (
     <>
@@ -45,11 +37,6 @@ const Menus = ({ char, setChar, emote, setEmote, costume,
         <CharacterMenu {...menuProps} />
         <EmotionMenu {...menuProps} />
         <CostumeMenu {...menuProps} />
-        {/* Only display the box menu if the character has multiple box options */}
-        {boxDirectory[char]
-          ? <BoxMenu {...menuProps} />
-          : null
-        }
       </div>
     </>
   );
