+++ components/TextAndTools.tsx
@@ -1,8 +1,8 @@
-import { useEffect } from 'react';
+import { useEffect, SyntheticEvent } from 'react';
 import Menus from './Menus';
 
-const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, setPortrait, text, setText, 
-  name, setName, font, setFont, boxFont, setBoxFont, box, setBox, selection, setSelection }) => {
+const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, setPortrait, setCustom, text, setText, 
+  name, setName, font, setFont, box, setBox }) => {
 
   const toolProps: any = {
     char,
@@ -14,16 +14,10 @@ const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, set
     setPortrait,
     box,
     setBox,
-    boxFont,
-    setBoxFont,
     font,
-    selection,
-    setSelection,
+    name,
+    setName,
   }
-
-  useEffect(() => {
-    // setBox(`../images/boxes/db-${selection}-${boxFont}.png`);
-  }, [boxFont]);
 
   const downloadImage = (): void => {
     const downloadCanvas: HTMLCanvasElement = document.createElement('canvas');
@@ -34,19 +28,49 @@ const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, set
       document.getElementById('portraitCanvas') as HTMLCanvasElement;
     const boxCanvas: HTMLCanvasElement = 
       document.getElementById('boxCanvas') as HTMLCanvasElement;
+    const tileCanvas: HTMLCanvasElement = 
+      document.getElementById('tileCanvas') as HTMLCanvasElement;
+    const nameCanvas: HTMLCanvasElement = 
+      document.getElementById('nameCanvas') as HTMLCanvasElement;
     const textCanvas: HTMLCanvasElement = 
       document.getElementById('textCanvas') as HTMLCanvasElement;
+
+    const nameField: HTMLTextAreaElement =
+      document.getElementById('nameField') as HTMLTextAreaElement;
     
     const dCtx: CanvasRenderingContext2D = downloadCanvas.getContext('2d') as CanvasRenderingContext2D;
     dCtx.drawImage(portraitCanvas, 0, 0, 1275, 500);
     dCtx.drawImage(boxCanvas, 0, 0, 1275, 500);
+    dCtx.drawImage(tileCanvas, 0, 0, 1275, 500);
+    dCtx.drawImage(nameCanvas, 0, 0, 1275, 500);
     dCtx.drawImage(textCanvas, 0, 0, 1275, 500);
     const link: HTMLAnchorElement = document.createElement('a');
-    link.download = `${char}-${text}.png`;
+    link.download = `${nameField.value}-${text}.png`;
     link.href = downloadCanvas.toDataURL('image/png');
     link.click();
     return;
   };
+
+  const customPortrait = (e: SyntheticEvent<HTMLInputElement>): void => {
+    // If an uploaded file exists...
+    if((e.target as HTMLInputElement).files[0]) {
+      // ...create a URL from the file and set it as our portrait image's source
+      setCustom(URL.createObjectURL((e.target as HTMLInputElement).files[0]));
+    }
+    return;
+  }
+
+  const changeName = (e: SyntheticEvent<HTMLTextAreaElement>): void => {
+    let currName = ((e.target as HTMLTextAreaElement).value);
+    let newName = '';
+    // Edge case: Prevents multiple whitespaces in a row so that
+    // random number selectors don't get caught in infinite loops
+    for (let i = 0; i < currName.length; i++) {
+      if (currName[i] === ' ' && currName[i + 1] === ' ') continue;
+      newName += currName[i];
+    }
+    setName(newName);
+  }
 
   return (
     <div id='textAndTools'>
@@ -59,7 +83,7 @@ const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, set
               rows={1}
               cols={45}
               defaultValue={name}
-              onChange={(e) => setName(e.target.value)}
+              onChange={(e) => changeName(e)}
             />
           </div>
           <div id='enterDialogue'>
@@ -73,20 +97,25 @@ const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, set
               onChange={(e) => setText(e.target.value)} 
             />
           </div>
-          <div 
-            id='download'
-            className='knife'
-            onClick={downloadImage}
-          >
-            <svg 
-              xmlns="http://www.w3.org/2000/svg"
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
+          <div id='downloadAndUpload'>
+            <div 
+              id='download'
+              className='knife'
+              onClick={downloadImage}
             >
-              <path d="M12 24l-8-9h6v-15h4v15h6z"/>
-            </svg>
-            <div>Download</div>    
+              <div>Download</div>    
+            </div>
+            <label id='upload' className='knife'>Upload Portrait
+              {/* Every time user clicks on the button, the old picture's URL is cleared, 
+              which allows the onChange event to fire even if you upload the same image twice */}
+              <input 
+                id='hiddenUpload' 
+                type='file' accept='image/*' 
+                onClick={(e) => (e.target as HTMLInputElement).value = null} 
+                onChange={(e) => customPortrait(e)}>
+              </input>
+              <div id='uploadSizeMessage'>(500px x 500px recommended)</div>
+            </label>
           </div>
         </div>
         <div id='genLinks'>
@@ -104,14 +133,14 @@ const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, set
             <div 
               id='vanillaFont' 
               className='fonts KRSMDivs knife' 
-              onClick={() => { setFont('KoreanKRSM'); setBoxFont('KoreanKRSM'); }}
+              onClick={() => { setFont('KoreanKRSM')}}
             >
               KoreanKRSM &#40;Persona 5&#41;
             </div>
             <div 
               id='royalFont' 
               className='fonts optimaDivs knife' 
-              onClick={() => { setFont('Optima nova LT'); setBoxFont('Optima nova LT'); }}
+              onClick={() => { setFont('Optima nova LT')}}
             >
               Optima Nova Black &#40;Persona 5 Royal&#41;
             </div>
@@ -122,8 +151,6 @@ const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, set
               lang='ja'
             >
               Slump DB &#40;ペルソナ5 - 日本語&#41;
-              <br></br>
-              <span id='nameNote'>&#40;Names Coming Soon!&#41;</span>
             </div>
             <div 
               id='koreanFont' 
@@ -132,8 +159,6 @@ const TextAndTools = ({ char, setChar, emote, setEmote, costume, setCostume, set
               lang='ko'
             >
               aCinema &#40;페르소나 5 - 한글어&#41;
-              <br></br>
-              <span id='nameNote'>&#40;Names Coming Soon!&#41;</span>
             </div>
           </div>
         </div>
