--- utils/findNameBox.ts
@@ -1,170 +1,0 @@
-const findNameBox = (name: string) => {
-  switch (name) {
-    case 'Akane':
-      return 'Akane';
-    case 'Akechi':
-      return 'Akechi';
-    case 'Akira':
-      return 'Akira'
-    case 'Alice':
-      return 'Alice'
-    case 'Ann':
-      return 'Ann'
-    case 'Beauty Thief':
-      return 'Beauty Thief'
-    case 'Caroline':
-      return 'Caroline'
-    case 'Chihaya':
-      return 'Chihaya'
-    case 'Ms. Chouno':
-      return 'Chouno'
-    case 'Cognitive Akechi':
-      return 'Cognitive Akechi'
-    case 'Cognitive Shiho Suzui':
-      return 'Cognitive Shiho'
-    case 'SIU Director':
-      return 'Director'
-    case 'Futaba':
-      return 'Futaba'
-    case 'Haru':
-      return 'Haru';
-    case 'Hifumi' :
-      return 'Hifumi';
-    case 'Coach Hiraguchi':
-      return 'Hiraguchi';
-    case 'Mr. Hiruta':
-      return 'Hiruta';
-    case 'Ichinose':
-      return 'Ichinose';
-    case 'Igor':
-      return 'Igor';
-    case 'Mr. Inui':
-      return 'Inui';
-    case 'Iwai':
-      return 'Iwai';
-    case 'Joker':
-      return 'Joker';
-    case 'Jose':
-      return 'Jose';
-    case 'Justine':
-      return 'Justine';
-    case 'Kaburagi':
-      return 'Kaburagi';
-    case 'Kamoshida':
-      return 'Kamoshida';
-    case 'Kaneshiro':
-      return 'Kaneshiro';
-    case 'Kasumi':
-      return 'Kasumi';
-    case 'Kawakami':
-      return 'Kawakami';
-    case 'Konoe':
-      return 'Konoe';
-    case 'Lavenza':
-      return 'Lavenza';
-    case 'Madarame':
-      return 'Madarame';
-    case 'Makoto':
-      return 'Makoto';
-    case 'Mariko':
-      return 'Mariko';
-    case 'Maruki':
-      return 'Maruki';
-    case 'Mika':
-      return 'Mika';
-    case 'Mishima':
-      return 'Mishima';
-    case 'Morgana':
-      return 'Morgana';
-    case 'Nakanohara':
-      return 'Nakanohara';
-    case 'Natsume':
-      return 'Natsume';
-    case 'Ohya':
-      return 'Ohya';
-    case 'Okumura':
-      return 'Okumura';
-    case 'Owada':
-      return 'Owada';
-    case 'Princess Ann':
-      return 'Princess Ann';
-    case 'Principal':
-      return 'Principal';
-    case 'Ren':
-      return 'Ren';
-    case 'Rumi':
-      return 'Rumi';
-    case 'Ryuji':
-      return 'Ryuji';
-    case 'Sae':
-      return 'Sae';
-    case 'Shadow Akane':
-      return 'Shadow Akane';
-    case 'Shadow Alice':
-      return 'Shadow Alice';
-    case 'Shadow Futaba':
-      return 'Shadow Futaba';
-    case 'Shadow Kamoshida':
-      return 'Shadow Kamoshida';
-    case 'Shadow Kaneshiro':
-      return 'Shadow Kaneshiro';
-    case 'Shadow Konoe':
-      return 'Shadow Konoe';
-    case 'Shadow Madarame':
-      return 'Shadow Madarame';
-    case 'Shadow Mariko':
-      return 'Shadow Mariko';
-    case 'Shadow Mishima':
-      return 'Shadow Mishima';
-    case 'Shadow Nakanohara':
-      return 'Shadow Nakanohara';
-    case 'Shadow Natsume':
-      return 'Shadow Natsume';
-    case 'Shadow Okumura':
-      return 'Shadow Okumura';
-    case 'Shadow Sae':
-      return 'Shadow Sae';
-    case 'Shadow Shido':
-      return 'Shadow Shido';
-    case 'Shady Commodities Tanaka':
-      return 'Shady Commodities Tanaka';
-    case 'Shibusawa':
-      return 'Shibusawa';
-    case 'Shido':
-      return 'Shido';
-    case 'Shiho':
-      return 'Shiho';
-    case 'Shinichi Yoshizawa':
-      return 'Shinichi';
-    case 'Shinya':
-      return 'Shinya';
-    case 'Sojiro':
-      return 'Sojiro';
-    case 'Sophia':
-      return 'Sophia';
-    case 'Sugimura':
-      return 'Sugimura';
-    case 'Sumire':
-      return 'Sumire';
-    case 'Takemi':
-      return 'Takemi';
-    case 'Tanaka':
-      return 'Tanaka';
-    case 'Ms. Usami':
-      return 'Usami';
-    case 'Mr. Ushimaru':
-      return 'Ushimaru';
-    case 'Wakaba':
-      return 'Wakaba';
-    case 'Yoshida':
-      return 'Yoshida';
-    case 'Yusuke':
-      return 'Yusuke';
-    case 'Zenkichi':
-      return 'Zenkichi';
-    default:
-      return null;
-  }
-}
-
-export { findNameBox };
